Added help menu

+++ LinuxSoftwareAutoInstall.sh
@@ -1,22 +1,17 @@
-#!/bin/sh
-
+#!/bin/bash
 #Need to run with root permissions
+#BUGS:
+    #Manoghud installation doesn't work properly even though it says so. Permission issue with the vm maybe?
 
-#TODO: Write a function to install the heroic launcher
-#TOOD: Write up a list of given changes and a confirmation prompt
-#TODO: Write some validation code to ensure the user didn't add more than 7 parameters
-#TODO: Write a for loop to loop through the InstallOptions and run the installation functions by inserting the name (Have to capitalize the first letter probably)
-
-InstallOptions=($0,$1,$2,$3,$4,$5,$6)
+InstallOptions=("$@")
 CurrentUser=`echo $USER`
 CurrentOS=`grep '^NAME' /etc/os-release` 
 CurrentOSReadable=`echo "$CurrentOS" | cut -d'"' -f 2`
 
-#TODO: Will need to add more options here. Which OS's use which package manager? Some documentation somewhere?
 case $CurrentOSReadable in 
     "Fedora Linux") 
         CurrentPackageManager="dnf" ;;
-    "Ubuntu")
+    "Ubuntu"|"Linux Mint"|"Debian")
         CurrentPackageManager="apt" ;; 
     "Arch Linux") 
         CurrentPackageManager="pacman" ;;
@@ -25,27 +20,30 @@ case $CurrentOSReadable in
 esac
 
 #Download a given tar file from github and extract to a given directory
-function FuncDownloadAndExtractTarball() {         
-    local loc="/home/$CurrentUser/Downloads/$1"
-    local repo="https://api.github.com/repos/$2/releases/latest"
-    local URLStrings=`curl -s "${repo}" | grep "browser_download_url" | cut -d '"' -f 4`
+function FuncDownloadAndExtractRepo() {         
+    local DownloadLocation="/home/$CurrentUser/Downloads/$1"
+    local RepoLocation="https://api.github.com/repos/$2/releases/latest"
+    local Filetype=$3
+    local URLStrings=`curl -s "${RepoLocation}" | grep "browser_download_url" | cut -d '"' -f 4`
     local URLArray=($URLStrings)
     local DesiredURL=''
 
     for t in "${URLArray[@]}"; do   
-        if [ "${t: -7}" == ".tar.gz" ]; then #Find tarball from releases
+        if [ "${t: -${#Filetype}}" == $Filetype ]; then     #Find file(s) from releases
             DesiredURL=$t
         fi 
     done
     
-    mkdir -p $loc
-    echo "Downloading files from '$DesiredURL' ..."
-    curl -sOL --output-dir $loc ${DesiredURL}
+    mkdir -p $DownloadLocation
+    echo "Downloading file(s) from '$DesiredURL' ..."
+    curl -sOL --output-dir $DownloadLocation ${DesiredURL}
 
-    local TarFileName="$(find $loc  -name "*.tar.gz")"
-    echo "Extracting files from '$TarFileName' ... "
-    tar xzf $TarFileName -C $loc
-    rm $TarFileName
+    if [ $Filetype = ".tar.gz" ]; then                      #File is an archive, need to extract
+        local TarFileName="$(find $DownloadLocation  -name "*$Filetype")"
+        echo "Extracting file(s) from '$TarFileName' ... "
+        tar xzf $TarFileName -C $DownloadLocation
+        rm $TarFileName
+    fi 
 }
 
 #Update the system 
@@ -55,7 +53,7 @@ function FuncUpdateSystem() {
     elif [ $CurrentPackageManager = "apt" ]; then 
         apt update -y && apt upgrade -y
     elif [ $CurrentPackageManager = "pacman" ]; then 
-        pacman -Syu 
+        pacman -Syu --noconfirm
     fi 
 }
 
@@ -66,7 +64,7 @@ function FuncInstallFlatpak() {
     elif [ $CurrentPackageManager = "apt" ]; then 
         apt install flatpak -y
     elif [ $CurrentPackageManager = "pacman" ]; then 
-        pacman -S flatpak -y
+        pacman -S flatpak --noconfirm
     fi 
     flatpak remote-add --if-not-exists flathub https://dl.flathub.org/repo/flathub.flatpakrepo
 }
@@ -78,7 +76,7 @@ function FuncInstallSteam() {
     elif [ $CurrentPackageManager = "apt" ]; then 
         apt install steam -y
     elif [ $CurrentPackageManager = "pacman" ]; then 
-        pacman -S steam -y 
+        pacman -S steam --noconfirm
     fi 
 }
 
@@ -86,10 +84,10 @@ function FuncInstallLutris() {
     if [ $CurrentPackageManager = "dnf" ]; then 
         dnf install lutris -y
     elif [ $CurrentPackageManager = "apt" ]; then 
-        add-apt-repository ppa:lutris-team/lutris
-        apt install lutris -y
+        add-apt-repository ppa:lutris-team/lutris -y
+        apt update && apt install lutris -y
     elif [ $CurrentPackageManager = "pacman" ]; then 
-        pacman -S lutris
+        pacman -S lutris --noconfirm
     fi 
 }
 
@@ -98,10 +96,10 @@ function FuncInstallDiscord() {
         dnf install https://mirrors.rpmfusion.org/nonfree/fedora/rpmfusion-nonfree-release-$(rpm -E %fedora).noarch.rpm -y
         dnf install discord -y
     elif [ $CurrentPackageManager = "apt" ]; then 
-        wget "https://discord.com/api/download?platform=linux&format=deb" -O discord.deb -y
+        wget "https://discord.com/api/download?platform=linux&format=deb" -O discord.deb
         apt install ./discord.deb -y
-    elif [ $CurrentPackageManager="pacman" ]; then 
-        pacman -S discord -y
+    elif [ $CurrentPackageManager = "pacman" ]; then 
+        pacman -S discord --noconfirm
     fi 
 }
 
@@ -109,9 +107,9 @@ function FuncInstallSignal() {
     if [ $CurrentPackageManager = "apt" ]; then 
         wget -O- https://updates.signal.org/desktop/apt/keys.asc | gpg --dearmor > signal-desktop-keyring.gpg
         cat signal-desktop-keyring.gpg | tee /usr/share/keyrings/signal-desktop-keyring.gpg > /dev/null
-        tee /etc/apt/sources.list.d/signal-xenial.list    
-        sudo apt install signal-desktop -y
-    else #Signal only officially supported on apt package manager, have to install the flatpak version for other pms
+        echo 'deb [arch=amd64 signed-by=/usr/share/keyrings/signal-desktop-keyring.gpg] https://updates.signal.org/desktop/apt xenial main' | tee /etc/apt/sources.list.d/signal-xenial.list
+        apt update && apt install signal-desktop
+    else    #Signal only officially supported on apt package manager, have to install the flatpak version otherwise
         InstallFlatpak
         flatpak install flathub org.signal.Signal -y
     fi 
@@ -120,34 +118,83 @@ function FuncInstallSignal() {
 function FuncInstallSpotify() {
     if [ $CurrentPackageManager = "apt" ]; then 
         curl -sS https://download.spotify.com/debian/pubkey_6224F9941A8AA6D1.gpg | gpg --dearmor --yes -o /etc/apt/trusted.gpg.d/spotify.gpg
-        tee /etc/apt/sources.list.d/spotify.list
-        apt-get install spotify-client -y
-    else  #Spotify only officially supported on apt package manager, have to install the flatpak version for other pms
+        echo "deb http://repository.spotify.com stable non-free" | sudo tee /etc/apt/sources.list.d/spotify.list
+        apt update && apt-get install spotify-client -y
+    else    #Spotify only officially supported on apt package manager, have to install the flatpak version otherwise
         InstallFlatpak
         flatpak install flathub com.spotify.Client -y
     fi 
 }
 
-function FuncInstallMangoHUD() {            #Download, Extract, and Install MangoHud using the releases 'mangohud-setup.sh' set up script
-    DownloadAndExtractTarball "MangoHud" "flightlessmango/MangoHud"
-    MangoHudFolder="/home/$CurrentUser/Downloads/MangoHud/MangoHud/"
+function FuncInstallHeroic() {
+    local HeroicDownloadDir="/home/$CurrentUser/Downloads/Heroic/"
+    if [ $CurrentPackageManager = "dnf" ]; then 
+        FuncDownloadAndExtractRepo "Heroic" "Heroic-Games-Launcher/HeroicGamesLauncher" ".rpm"
+        local HeroicRpmFile=`basename "$(find $HeroicDownloadDir -name "heroic-*.x86_64.rpm")"`
+        cd $HeroicDownloadDir
+        dnf install ./$HeroicRpmFile -y
+    elif [ $CurrentPackageManager = "apt" ]; then 
+        FuncDownloadAndExtractRepo "Heroic" "Heroic-Games-Launcher/HeroicGamesLauncher" ".deb"
+        local HeroicDebFile=`basename "$(find $HeroicDownloadDir -name "heroic_*_amd64.deb")"`
+        cd $HeroicDownloadDir
+        dpkg -i $HeroicDebFile
+    elif [ $CurrentPackageManager = "pacman" ]; then 
+        pacman -S heroic-games-launcher-bin --noconfirm
+    fi 
+}
+
+function FuncInstallMangohud() {            #Download, Extract, and Install MangoHud using the releases 'mangohud-setup.sh' set up script
+    FuncDownloadAndExtractRepo "MangoHud" "flightlessmango/MangoHud" ".tar.gz"
+    local MangoHudFolder="/home/$CurrentUser/Downloads/MangoHud/MangoHud/"
     cd $MangoHudFolder
     ./mangohud-setup.sh install
 }
 
-function FuncInstallGEProton() {            #Download, Extract, and Install GE-Proton to the default (non-flatpak) compatibility folder in the steam directory
-    DownloadAndExtractTarball "Proton" "GloriousEggroll/proton-ge-custom"
-    ProtonDownloadFolder="$(find "/home/$CurrentUser/Downloads/Proton/" -name "GE-Proton*")"
-    ProtonSteamFolder="/home/$CurrentUser/.steam/steam/compatibilitytools.d"
+function FuncInstallProtonge() {            #Download, Extract, and Install GE-Proton to the default (non-flatpak) compatibility folder in the steam directory
+    FuncDownloadAndExtractRepo "Proton" "GloriousEggroll/proton-ge-custom" ".tar.gz"
+    local ProtonDownloadFolder="$(find "/home/$CurrentUser/Downloads/Proton/" -name "GE-Proton*")"
+    local ProtonSteamFolder="/home/$CurrentUser/.steam/steam/compatibilitytools.d"
     mkdir -p $ProtonSteamFolder
     cp -r $ProtonDownloadFolder $ProtonSteamFolder
 }
 
-#UpdateSystem
-#InstallSteam
-#InstallMangoHUD 
-#InstallGEProton
-#InstallLutris
-#InstallDiscord
-#InstallSignal
-#InstallSpotify
+if [ $# -eq 0 ]; then 
+    echo "No arguments supplied; Use 'help' for details"
+elif [ "${InstallOptions[0]//-/}" = "help" ]; then 
+    Output="
+        Usage: $(basename $0) [OPTIONS]\n
+        Permission requiremnts: root\n
+        Supports: Ubuntu, Fedora, Arch\n
+        \n
+        Options:\n
+        help\t\t        Show available install options\n
+        steam\t\t       Installs the Steam client\n
+        lutris\t\t      Installs the Lutris client\n
+        heroic\t\t      Installs the Heroic client\n
+        discord\t       Installs the Discord client\n
+        signal\t\t      Installs the Signal client (Flatpak version if no apt)\n
+        spotify\t       Installs the Spotify client (Flatpak version if no apt)\n
+        mangohud\t      Installs the latest MangoHUD release via the repo install script\n
+        protonge\t      Installs the latest Glorious Eggroll Proton release and moves it to the steam compatibility folder (Assuming existing steam install)
+        "
+    echo -e $Output
+else 
+    ParsedInstallOptions=() 
+    for Options in "${InstallOptions[@]}"; do               #Sanitize all '-' values from all parameters
+        ParsedInstallOptions+=("${Options//-/}")
+    done
+
+    ValidPrograms=("Steam","Lutris","Heroic","Discord","Signal","Spotify","Mangohud","Protonge")
+    for Options in "${ParsedInstallOptions[@]}"; do         #Check that passed parameters are valid, so they can safetly be used for function calls
+        if ! [[ $(echo "${ValidPrograms[@]}" | grep -F -w "${Options^}") ]]; then     
+            echo "Parameter '$Options' is not a valid install option; See 'help' for details"
+            exit 1
+        fi 
+    done
+
+    FuncUpdateSystem
+    for Options in "${ParsedInstallOptions[@]}"; do         #If we made it here, all looks good. Run desired installations
+        echo "Installing $Options ... "
+        FuncInstall${Options^}
+    done
+fi
